Tester routage pages quasi-statiques /stats, /signin, /signout
