--- test/routes/bonjourMonde.test.ts
@@ -1,24 +1,0 @@
-import supertest from 'supertest';
-import 'jest-extended';
-import app from '../../src/app';
-
-const request = supertest(app);
-
-const testNom1 = 'Jean-Marc';
-
-describe('baseRoute', () => {
-
-  it('devrait avoir un contenu HTML', async () => {
-    const response = await request.get('/');
-    expect(response.status).toBe(200);
-    expect(response.type).toBe("text/html");
-  });
-
-});
-
-describe('GET /bo/gu/s/URL/', () => {
-  it(`devrait répondre avec une mauvaise demande lorsque l'URL est mauvais`, async () => {
-    const response = await request.get('/bo/gu/s/URL/' + testNom1);
-    expect(response.status).toBe(404);
-  });
-});
